requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
argon2-cffi
python-dotenv
